pyproject.toml
[tool.poetry]
name = "twitchchat_display"
version = "0.1.0"
description = "PyGame based app to render twitch chat in real time"
authors = ["Samantha Hughes <shughes.uk@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
pygame = "^1.9.6"
pyyaml = "^5.3.1"
pillow = "^7.1.2"
requests = "^2.23.0"
click = "^7.1.2"
fonttools = "^4.9.0"
webcolors = "^1.11.1"
colorlog = "^4.1.0"
twitchchat = {git = "https://github.com/shughes-uk/python-twitchchat.git"}

[tool.poetry.scripts]
twitch_display = 'twitchchat_display.main:main'